requirements.txt
Flask==2.1.2
pymongo==4.4.1
Werkzeug==2.2.2
scikit-learn==1.2.2
python-dotenv==1.0.0